🛡️ Sentinel: [security improvement] (#2587)
* 🛡️ Sentinel: [security improvement] Enabled Docker scanning

Enabled Docker vulnerability scanning in CI by removing the `.disabled` extension from the `docker-security-scan.yml` GitHub Actions workflow.


* chore: trigger CI

* chore: trigger CI

* 🛡️ Sentinel: [security improvement] Enabled Docker scanning

Enabled Docker vulnerability scanning in CI by removing the `.disabled` extension from the `docker-security-scan.yml` GitHub Actions workflow.

Also fixed a test isolation issue in `_install_fake_simulation_modules` that caused CI failures.


* 🛡️ Sentinel: [security improvement] Enabled Docker scanning

Enabled Docker vulnerability scanning in CI by removing the `.disabled` extension from the `docker-security-scan.yml` GitHub Actions workflow.

Also fixed test mock isolation issues in `_install_fake_simulation_modules` and removed testing of non-existent helper functions from `test_panel_builders_helpers.py`.


* 🛡️ Sentinel: [security improvement] Enabled Docker scanning

Enabled Docker vulnerability scanning in CI by removing the `.disabled` extension from the `docker-security-scan.yml` GitHub Actions workflow.

Also fixed test mock isolation issues in `_install_fake_simulation_modules` and removed testing of non-existent helper functions from `test_panel_builders_helpers.py`. Fixed `PYTHONPATH` missing from quality check scripts in CI workflow.


* 🛡️ Sentinel: [security improvement] Enabled Docker scanning

Enabled Docker vulnerability scanning in CI by removing the `.disabled` extension from the `docker-security-scan.yml` GitHub Actions workflow.

Also fixed test mock isolation issues in `_install_fake_simulation_modules` and removed testing of non-existent helper functions from `test_panel_builders_helpers.py`. Fixed `PYTHONPATH` missing from quality check scripts in CI workflow.


---------
