+++ SPEC.md
@@ -27,7 +27,7 @@
 | **Primary Language(s)** | Python 3.10+, Rust, TypeScript                     |
 | **License**             | MIT                                                |
 | **Current Version**     | 2.1.0                                              |
-| **Spec Version**        | 1.0.86                                             |
+| **Spec Version**        | 1.0.77                                         |
 | **Last Spec Update**    | 2026-04-10                                         |
 
 ## 2. Purpose & Mission
@@ -493,16 +493,7 @@ pytest tests/ --cov=src --cov-fail-under=70
 
 | Date       | Version | Changes                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               |
 | ---------- | ------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| 2026-04-10 | 1.0.86  | Optional-stack contracts(Pinocchio): preserved root-body MuJoCo joints during URDF export so Pinocchio models retain the MuJoCo velocity-DOF contract, rejected mock-only Pinocchio availability surfaces, and isolated induced-acceleration unit mocks from global `sys.modules` leakage. |
-| 2026-04-10 | 1.0.85  | Optional-stack contracts(Drake): tightened Drake availability probing to reject mock-only or partial `pydrake` surfaces, restored module-scoped Drake dependency mocks for module-scoped tests, and aligned the isolated strict reset test with the initialized-engine precondition. |
-| 2026-04-10 | 1.0.84  | Optional-stack contracts: tightened OpenSim availability probing to reject importable-but-incompatible bindings, exported unlimited MuJoCo hinge joints as URDF `continuous` joints, and aligned Pinocchio/OpenSim audit tests with the implemented engine contracts. |
-| 2026-04-10 | 1.0.83  | CI governance: removed the `**.md` pull-request ignore from CI Standard so required quality-gate checks run for SPEC-only pull requests instead of leaving required checks permanently pending. |
-| 2026-04-10 | 1.0.82  | Refactor(data-fitting): decomposed `validation_pkg/data_fitting.py` into focused data-model, inverse-kinematics, parameter-estimation, sensitivity, and pipeline helper modules while preserving the legacy facade import surface and adding facade regression coverage. |
-| 2026-04-10 | 1.0.81  | Test framework fix: finalized the pendulum panel-builder helper isolation by reloading `panel_builders` under fixture-scoped fake simulation and perturbation modules, preserving richer fake `run_simulation` results, and removing stale merge-conflict artifacts that broke linting on the branch. |
-| 2026-04-10 | 1.0.80  | Test isolation(Drake): expanded the isolated Drake strict-test mock surface to include `pydrake.geometry` and `pydrake.multibody.tree`, reducing cross-test import failures when the optional-stack lane executes strict Drake mocks before the broader Drake wrapper and integration-audit suites. |
-| 2026-04-10 | 1.0.79  | Test framework fix: tightened pendulum panel-builder helper tests so fake simulation and perturbation modules are fixture-scoped, reload `panel_builders` under the fake environment, and avoid leaking mocked modules or incompatible `run_simulation` signatures into unrelated tests. |
-| 2026-04-10 | 1.0.78  | Test framework fix: Refactored `test_panel_builders_helpers.py` to prevent module-level mocking of `run_simulation` from polluting `sys.modules` and causing cascading failures in other test suites. |
-| 2026-04-10 | 1.0.77  | Optimization: Cached difference vectors and their np.linalg.norm results into local variables inside tight collision-checking loops to eliminate redundant vector subtractions and matrix math operations, halving execution time in these hot paths without sacrificing code readability. |
+| 2026-04-10 | 1.0.77  | Enabled Docker vulnerability scanning (Trivy) in CI pipeline. |
 | 2026-04-10 | 1.0.76  | Simscape + GUI maintenance: fixed the 3D golf dataset generator defaults to target the bundled `GolfSwing3D_Kinetic` model while keeping legacy `verbose` and newer `verbosity` config paths compatible, and decomposed the launcher dashboard plus pendulum GUI builder/toolstrip code into smaller helpers with focused regression coverage. |
 | 2026-04-10 | 1.0.75  | Optimization: Replaced np.linalg.norm(..., axis=1) with explicit element-wise computation in deformable objects to avoid axis reduction overhead for arrays with small inner dimensions. |
 | 2026-04-10 | 1.0.74  | API/auth integrity: aligned API-key prefix lookup and API-key creation on the canonical `key_prefix` field, removed the stale auth integration xfail, and added a route-level regression test to ensure newly created API keys persist the hashed prefix used by the fast-path lookup dependency.                                                                                                                                                                                                        |
