+++ src/robotics/planning/collision/collision_checker.py
@@ -383,10 +383,8 @@ class CollisionChecker:
                     closest_pair = pair
                     point_a = pa
                     point_b = pb
-                    diff = pb - pa
-                    norm = np.linalg.norm(diff)
-                    if norm > 1e-10:
-                        normal = diff / norm
+                    if np.linalg.norm(pb - pa) > 1e-10:
+                        normal = (pb - pa) / np.linalg.norm(pb - pa)
 
                     # Early exit if max_distance exceeded
                     if min_distance > query.max_distance:
@@ -408,10 +406,8 @@ class CollisionChecker:
                         closest_pair = env_pair
                         point_a = pa
                         point_b = pb
-                        diff = pb - pa
-                        norm = np.linalg.norm(diff)
-                        if norm > 1e-10:
-                            normal = diff / norm
+                        if np.linalg.norm(pb - pa) > 1e-10:
+                            normal = (pb - pa) / np.linalg.norm(pb - pa)
 
             computation_time = time.perf_counter() - start_time
