+++ src/shared/python/engine_core/engine_availability.py
@@ -37,7 +37,6 @@ import sys
 from collections.abc import Callable
 from enum import Enum
 from typing import Any, TypeVar
-from unittest.mock import Mock
 
 logger = logging.getLogger(__name__)
 
@@ -50,52 +49,6 @@ class EngineStatus(Enum):
 
 _engine_status_cache: dict[str, EngineStatus] = {}
 _engine_error_cache: dict[str, Exception] = {}
-
-
-def _validate_opensim_bindings(opensim_module: Any) -> None:
-    """Reject importable but incompatible OpenSim Python packages."""
-    required_attrs = ("Model", "Manager", "Vector", "InverseDynamicsSolver")
-    missing_attrs = [
-        attr for attr in required_attrs if not hasattr(opensim_module, attr)
-    ]
-    if missing_attrs:
-        raise AttributeError(
-            "OpenSim package is missing required bindings: " + ", ".join(missing_attrs)
-        )
-
-    try:
-        opensim_module.Model("__upstream_drift_missing_probe__.osim")
-    except TypeError as exc:
-        raise TypeError("OpenSim Model binding does not accept model paths") from exc
-    except Exception:
-        # Real OpenSim bindings may reject the missing probe file, but reaching
-        # that runtime error proves the path-taking constructor exists.
-        pass
-
-
-def _validate_drake_bindings(drake_module: Any) -> None:
-    """Reject partial or mock-only pydrake surfaces."""
-    if isinstance(drake_module, Mock):
-        raise TypeError("pydrake.all resolved to a mock module")
-
-    for module_name in (
-        "pydrake.geometry",
-        "pydrake.math",
-        "pydrake.multibody.tree",
-    ):
-        module = importlib.import_module(module_name)
-        if isinstance(module, Mock):
-            raise TypeError(f"{module_name} resolved to a mock module")
-
-
-def _validate_pinocchio_bindings(pinocchio_module: Any) -> None:
-    """Reject mock-only or wrong-package Pinocchio surfaces."""
-    if isinstance(pinocchio_module, Mock):
-        raise TypeError("pinocchio resolved to a mock module")
-    if not hasattr(pinocchio_module, "buildModelFromUrdf"):
-        raise ImportError(
-            "Incorrect pinocchio package (likely nose plugin). Please install pinocchio from conda-forge."
-        )
 
 
 def _probe_engine(
@@ -127,7 +80,7 @@ def _probe_engine(
 
     try:
         if import_name == "drake":
-            _validate_drake_bindings(importlib.import_module("pydrake.all"))
+            importlib.import_module("pydrake.all")
         elif import_name == "torch":
             importlib.import_module("torch")
         elif import_name == "tf":
@@ -147,9 +100,11 @@ def _probe_engine(
         elif import_name == "pyside6":
             importlib.import_module("PySide6.QtWidgets")
         elif import_name == "pinocchio":
-            _validate_pinocchio_bindings(importlib.import_module("pinocchio"))
-        elif import_name == "opensim":
-            _validate_opensim_bindings(importlib.import_module("opensim"))
+            pin = importlib.import_module("pinocchio")
+            if not hasattr(pin, "buildModelFromUrdf"):
+                raise ImportError(
+                    "Incorrect pinocchio package (likely nose plugin). Please install pinocchio from conda-forge."
+                )
         else:
             importlib.import_module(import_name)
