+++ tests/unit/pendulum_simulator/test_panel_builders_helpers.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import math
 import sys
-from importlib import import_module, reload
 from types import ModuleType
 
 import numpy as np
@@ -13,11 +12,10 @@ import pytest
 pytest.importorskip("PyQt6")
 
 
-def _make_fake_simulation_modules() -> dict[str, ModuleType]:
+def _install_fake_simulation_modules() -> None:
     class _FakeResultBase:
         pass
 
-    modules: dict[str, ModuleType] = {}
     for module_name in (
         "src.shared.python.pendulum_simulator.simulation",
         "src.shared.python.pendulum_simulator.simulation_triple",
@@ -25,22 +23,20 @@ def _make_fake_simulation_modules() -> dict[str, ModuleType]:
     ):
         module = ModuleType(module_name)
         module.make_polynomial_torque = lambda *args, **kwargs: (args, kwargs)
-        module.run_simulation = lambda *args, **kwargs: {"args": args, "kwargs": kwargs}
+        module.run_simulation = lambda *args, **kwargs: kwargs
         if module_name.endswith(".simulation"):
             module.SimulationResult = _FakeResultBase
         elif module_name.endswith(".simulation_triple"):
             module.TripleSimulationResult = _FakeResultBase
         else:
             module.GolferSimulationResult = _FakeResultBase
-        modules[module_name] = module
-    return modules
+        sys.modules[module_name] = module
 
 
-def _make_fake_perturbation_modules() -> dict[str, ModuleType]:
-    modules: dict[str, ModuleType] = {}
+def _install_fake_perturbation_modules() -> None:
     package = ModuleType("src.shared.python.pendulum_simulator.perturbation")
     package.__path__ = []  # type: ignore[attr-defined]
-    modules["src.shared.python.pendulum_simulator.perturbation"] = package
+    sys.modules["src.shared.python.pendulum_simulator.perturbation"] = package
 
     config_module = ModuleType(
         "src.shared.python.pendulum_simulator.perturbation.config"
@@ -54,7 +50,9 @@ def _make_fake_perturbation_modules() -> dict[str, ModuleType]:
 
     config_module.PerturbationConfig = _FakePerturbationConfig
     config_module.PerturbationSummary = _FakePerturbationSummary
-    modules["src.shared.python.pendulum_simulator.perturbation.config"] = config_module
+    sys.modules["src.shared.python.pendulum_simulator.perturbation.config"] = (
+        config_module
+    )
 
     analyzer_module = ModuleType(
         "src.shared.python.pendulum_simulator.pendulum_perturbation_analyzer"
@@ -71,30 +69,23 @@ def _make_fake_perturbation_modules() -> dict[str, ModuleType]:
             return _FakePerturbationSummary()
 
     analyzer_module.PendulumPerturbationAnalyzer = _FakeAnalyzer
-    modules["src.shared.python.pendulum_simulator.pendulum_perturbation_analyzer"] = (
-        analyzer_module
-    )
+    sys.modules[
+        "src.shared.python.pendulum_simulator.pendulum_perturbation_analyzer"
+    ] = analyzer_module
 
     perturbation_analysis_module = ModuleType(
         "src.shared.python.pendulum_simulator.perturbation_analysis"
     )
     perturbation_analysis_module.variability_summary = lambda *args, **kwargs: None
-    modules["src.shared.python.pendulum_simulator.perturbation_analysis"] = (
+    sys.modules["src.shared.python.pendulum_simulator.perturbation_analysis"] = (
         perturbation_analysis_module
     )
-    return modules
 
 
-@pytest.fixture
-def panel_builders(monkeypatch: pytest.MonkeyPatch):
-    for module_name, module in {
-        **_make_fake_simulation_modules(),
-        **_make_fake_perturbation_modules(),
-    }.items():
-        monkeypatch.setitem(sys.modules, module_name, module)
+_install_fake_simulation_modules()
+_install_fake_perturbation_modules()
 
-    module = import_module("src.shared.python.pendulum_simulator.gui.panel_builders")
-    return reload(module)
+from src.shared.python.pendulum_simulator.gui import panel_builders  # noqa: E402
 
 
 class _FakeResult:
@@ -172,7 +163,7 @@ class _FakeOptimizationWidget:
         self.n_torque_params = n_torque_params
 
 
-def test_helper_parsers_and_motion_extraction(panel_builders) -> None:
+def test_helper_parsers_and_motion_extraction() -> None:
     assert panel_builders._parse_coefficients("1.0, 2.5, , 3") == [1.0, 2.5, 3.0]
     assert panel_builders._parse_coefficients("") == [0.0]
     assert panel_builders._chunk_coefficients(np.array([1, 2, 3, 4, 5]), 2) == [
@@ -200,7 +191,6 @@ def test_helper_parsers_and_motion_extraction(panel_builders) -> None:
 
 
 def test_build_double_panel_wires_helper_callbacks(
-    panel_builders,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     fake_pendulum = _FakePendulum()
